+++ src/widgets/editors/editorcfgqml.h
@@ -12,6 +12,9 @@ class EditorCfgQml : public QObject {
   Q_PROPERTY(bool expandTab READ expandTab NOTIFY changed)
   Q_PROPERTY(bool lineNumber READ lineNumber NOTIFY changed)
   Q_PROPERTY(bool highlightCurrentLine READ highlightCurrentLine NOTIFY changed)
+  Q_PROPERTY(bool autoIndent READ autoIndent NOTIFY changed)
+  Q_PROPERTY(bool continueList READ continueList NOTIFY changed)
+  Q_PROPERTY(bool autoPair READ autoPair NOTIFY changed)
 public:
   explicit EditorCfgQml(QObject *p_parent = nullptr);
 
@@ -20,6 +23,9 @@ public:
   bool expandTab() const;
   bool lineNumber() const;
   bool highlightCurrentLine() const;
+  bool autoIndent() const;
+  bool continueList() const;
+  bool autoPair() const;
 
 signals:
   void changed();